bug fixes and performance improvement, and more

+++ src/main/java/com/verr1/controlcraft/content/cctweaked/peripheral/FlapBearingPeripheral.java
@@ -24,7 +24,7 @@ public class FlapBearingPeripheral extends AbstractAttachedPeripheral<FlapBearin
     }
 
     @LuaFunction
-    public final float getAngle(){
+    public final double getAngle(){
         return getTarget().getAngle();
     }
